Refactor exceptions.
